add pytorch-v1.1 branch

+++ tools/train.py
@@ -21,6 +21,7 @@ import torch
 import torch.nn as nn
 import torch.backends.cudnn as cudnn
 import torch.optim
+from torch.utils.data.distributed import DistributedSampler
 from tensorboardX import SummaryWriter
 
 import _init_paths
@@ -31,7 +32,7 @@ from config import update_config
 from core.criterion import CrossEntropy, OhemCrossEntropy
 from core.function import train, validate
 from utils.modelsummary import get_model_summary
-from utils.utils import create_logger, FullModel
+from utils.utils import create_logger, FullModel, get_rank
 
 def parse_args():
     parser = argparse.ArgumentParser(description='Train segmentation network')
@@ -40,6 +41,7 @@ def parse_args():
                         help='experiment configure file name',
                         required=True,
                         type=str)
+    parser.add_argument("--local_rank", type=int, default=0)
     parser.add_argument('opts',
                         help="Modify config options using the command-line",
                         default=None,
@@ -70,22 +72,32 @@ def main():
     cudnn.deterministic = config.CUDNN.DETERMINISTIC
     cudnn.enabled = config.CUDNN.ENABLED
     gpus = list(config.GPUS)
+    distributed = len(gpus) > 1
+    device = torch.device('cuda:{}'.format(args.local_rank))
 
     # build model
     model = eval('models.'+config.MODEL.NAME +
                  '.get_seg_model')(config)
 
-    dump_input = torch.rand(
-        (1, 3, config.TRAIN.IMAGE_SIZE[1], config.TRAIN.IMAGE_SIZE[0])
-    )
-    logger.info(get_model_summary(model.cuda(), dump_input.cuda()))
+    if args.local_rank == 0:
+        # provide the summary of model
+        dump_input = torch.rand(
+            (1, 3, config.TRAIN.IMAGE_SIZE[1], config.TRAIN.IMAGE_SIZE[0])
+            )
+        logger.info(get_model_summary(model.to(device), dump_input.to(device)))
 
-    # copy model file
-    this_dir = os.path.dirname(__file__)
-    models_dst_dir = os.path.join(final_output_dir, 'models')
-    if os.path.exists(models_dst_dir):
-        shutil.rmtree(models_dst_dir)
-    shutil.copytree(os.path.join(this_dir, '../lib/models'), models_dst_dir)
+        # copy model file
+        this_dir = os.path.dirname(__file__)
+        models_dst_dir = os.path.join(final_output_dir, 'models')
+        if os.path.exists(models_dst_dir):
+            shutil.rmtree(models_dst_dir)
+        shutil.copytree(os.path.join(this_dir, '../lib/models'), models_dst_dir)
+
+    if distributed:
+        torch.cuda.set_device(args.local_rank)
+        torch.distributed.init_process_group(
+            backend="nccl", init_method="env://",
+        )
 
     # prepare data
     crop_size = (config.TRAIN.IMAGE_SIZE[1], config.TRAIN.IMAGE_SIZE[0])
@@ -102,13 +114,19 @@ def main():
                         downsample_rate=config.TRAIN.DOWNSAMPLERATE,
                         scale_factor=config.TRAIN.SCALE_FACTOR)
 
+    if distributed:
+        train_sampler = DistributedSampler(train_dataset)
+    else:
+        train_sampler = None
+
     trainloader = torch.utils.data.DataLoader(
         train_dataset,
-        batch_size=config.TRAIN.BATCH_SIZE_PER_GPU*len(gpus),
-        shuffle=config.TRAIN.SHUFFLE,
+        batch_size=config.TRAIN.BATCH_SIZE_PER_GPU,
+        shuffle=config.TRAIN.SHUFFLE and train_sampler is None,
         num_workers=config.WORKERS,
         pin_memory=True,
-        drop_last=True)
+        drop_last=True,
+        sampler=train_sampler)
 
     if config.DATASET.EXTRA_TRAIN_SET:
         extra_train_dataset = eval('datasets.'+config.DATASET.DATASET)(
@@ -124,13 +142,19 @@ def main():
                     downsample_rate=config.TRAIN.DOWNSAMPLERATE,
                     scale_factor=config.TRAIN.SCALE_FACTOR)
 
+        if distributed:
+            extra_train_sampler = DistributedSampler(extra_train_dataset)
+        else:
+            extra_train_sampler = None
+
         extra_trainloader = torch.utils.data.DataLoader(
             extra_train_dataset,
-            batch_size=config.TRAIN.BATCH_SIZE_PER_GPU*len(gpus),
-            shuffle=config.TRAIN.SHUFFLE,
+            batch_size=config.TRAIN.BATCH_SIZE_PER_GPU,
+            shuffle=config.TRAIN.SHUFFLE and extra_train_sampler is None,
             num_workers=config.WORKERS,
             pin_memory=True,
-            drop_last=True)
+            drop_last=True,
+            sampler=extra_train_sampler)
 
     test_size = (config.TEST.IMAGE_SIZE[1], config.TEST.IMAGE_SIZE[0])
     test_dataset = eval('datasets.'+config.DATASET.DATASET)(
@@ -143,14 +167,21 @@ def main():
                         ignore_label=config.TRAIN.IGNORE_LABEL,
                         base_size=config.TEST.BASE_SIZE,
                         crop_size=test_size,
+                        center_crop_test=config.TEST.CENTER_CROP_TEST,
                         downsample_rate=1)
+
+    if distributed:
+        test_sampler = DistributedSampler(test_dataset)
+    else:
+        test_sampler = None
 
     testloader = torch.utils.data.DataLoader(
         test_dataset,
-        batch_size=config.TEST.BATCH_SIZE_PER_GPU*len(gpus),
+        batch_size=config.TEST.BATCH_SIZE_PER_GPU,
         shuffle=False,
         num_workers=config.WORKERS,
-        pin_memory=True)
+        pin_memory=True,
+        sampler=test_sampler)
 
     # criterion
     if config.LOSS.USE_OHEM:
@@ -163,7 +194,10 @@ def main():
                                  weight=train_dataset.class_weights)
 
     model = FullModel(model, criterion)
-    model = nn.DataParallel(model, device_ids=gpus).cuda()
+    model = nn.SyncBatchNorm.convert_sync_batchnorm(model)
+    model = model.to(device)
+    model = nn.parallel.DistributedDataParallel(
+        model, device_ids=[args.local_rank], output_device=args.local_rank)
 
     # optimizer
     if config.TRAIN.OPTIMIZER == 'sgd':
@@ -187,7 +221,8 @@ def main():
         model_state_file = os.path.join(final_output_dir,
                                         'checkpoint.pth.tar')
         if os.path.isfile(model_state_file):
-            checkpoint = torch.load(model_state_file)
+            checkpoint = torch.load(model_state_file, 
+                        map_location=lambda storage, loc: storage)
             best_mIoU = checkpoint['best_mIoU']
             last_epoch = checkpoint['epoch']
             model.module.load_state_dict(checkpoint['state_dict'])
@@ -201,42 +236,50 @@ def main():
     extra_iters = config.TRAIN.EXTRA_EPOCH * epoch_iters
     
     for epoch in range(last_epoch, end_epoch):
+        if distributed:
+            train_sampler.set_epoch(epoch)
         if epoch >= config.TRAIN.END_EPOCH:
             train(config, epoch-config.TRAIN.END_EPOCH, 
                   config.TRAIN.EXTRA_EPOCH, epoch_iters, 
                   config.TRAIN.EXTRA_LR, extra_iters, 
-                  extra_trainloader, optimizer, model, writer_dict)
+                  extra_trainloader, optimizer, model, 
+                  writer_dict, device)
         else:
             train(config, epoch, config.TRAIN.END_EPOCH, 
                   epoch_iters, config.TRAIN.LR, num_iters,
-                  trainloader, optimizer, model, writer_dict)
+                  trainloader, optimizer, model, writer_dict,
+                  device)
 
-        logger.info('=> saving checkpoint to {}'.format(
-            final_output_dir + 'checkpoint.pth.tar'))
-        torch.save({
-            'epoch': epoch+1,
-            'best_mIoU': best_mIoU,
-            'state_dict': model.module.state_dict(),
-            'optimizer': optimizer.state_dict(),
-        }, os.path.join(final_output_dir,'checkpoint.pth.tar'))
-        valid_loss, mean_IoU, IoU_array = validate(
-                        config, testloader, model, writer_dict)
-        if mean_IoU > best_mIoU:
-            best_mIoU = mean_IoU
-            torch.save(model.module.state_dict(),
-                       os.path.join(final_output_dir, 'best.pth'))
-        msg = 'Loss: {:.3f}, MeanIU: {: 4.4f}, Best_mIoU: {: 4.4f}'.format(
+        valid_loss, mean_IoU, IoU_array = validate(config, 
+                    testloader, model, writer_dict, device)
+
+        if args.local_rank == 0:
+            logger.info('=> saving checkpoint to {}'.format(
+                final_output_dir + 'checkpoint.pth.tar'))
+            torch.save({
+                'epoch': epoch+1,
+                'best_mIoU': best_mIoU,
+                'state_dict': model.module.state_dict(),
+                'optimizer': optimizer.state_dict(),
+            }, os.path.join(final_output_dir,'checkpoint.pth.tar'))
+
+            if mean_IoU > best_mIoU:
+                best_mIoU = mean_IoU
+                torch.save(model.module.state_dict(),
+                           os.path.join(final_output_dir, 'best.pth'))
+            msg = 'Loss: {:.3f}, MeanIU: {: 4.4f}, Best_mIoU: {: 4.4f}'.format(
                     valid_loss, mean_IoU, best_mIoU)
-        logging.info(msg)
-        logging.info(IoU_array)
+            logging.info(msg)
+            logging.info(IoU_array)
 
-    torch.save(model.module.state_dict(),
-               os.path.join(final_output_dir, 'final_state.pth'))
+            if epoch == end_epoch - 1:
+                torch.save(model.module.state_dict(),
+                       os.path.join(final_output_dir, 'final_state.pth'))
 
-    writer_dict['writer'].close()
-    end = timeit.default_timer()
-    logger.info('Hours: %d' % np.int((end-start)/3600))
-    logger.info('Done')
+                writer_dict['writer'].close()
+                end = timeit.default_timer()
+                logger.info('Hours: %d' % np.int((end-start)/3600))
+                logger.info('Done')
 
 
 if __name__ == '__main__':
